Include every standard header used, audited across the whole tree
The headless job failed on Settings.hpp using std::optional without
including <optional>. libc++ provides it transitively; libstdc++ does
not. That file is from M1c — the audit I did before pushing M2 only
covered the files I had just written, which is exactly the wrong scope
for a portability bug.

Rather than fix the one file, this resolves every std:: name used across
core, codecs, tools and tests against the transitive closure of includes
reachable through *our own* headers. Anything only reachable through a
libc++ system header is a portability bug by construction. 25 files had
one; all are fixed and the audit now reports zero.

Also drops a Database* that sql::Statement stored and never read, and
the last -Wdouble-promotion warnings in the tests. Both builds are now
warning-free.

+++ tests/codecs/test_conformance.cpp
@@ -169,7 +169,7 @@ std::vector<ChannelStats> analyse(const std::vector<float>& interleaved,
         float         previous   = 0.0F;
         for (std::size_t i = begin; i < end; ++i) {
             const float v = interleaved[i * channels + c];
-            sumSquares += static_cast<double>(v) * v;
+            sumSquares += static_cast<double>(v) * static_cast<double>(v);
             if (i > begin && ((previous < 0.0F) != (v < 0.0F))) {
                 ++crossings;
             }
@@ -314,7 +314,8 @@ TEST_CASE("seeking lands on the right audio, not just the right frame",
 
         double worst = 0.0;
         for (std::size_t i = 0; i < compare; ++i) {
-            worst = std::max(worst, std::abs(static_cast<double>(afterSeek[i]) - expected[i]));
+            worst = std::max(worst, std::abs(static_cast<double>(afterSeek[i]) -
+                                             static_cast<double>(expected[i])));
         }
 
         if (codec.lossless) {
